+++ Tests/SwiftStdlibTests/CharacterExtensionsTests.swift
@@ -19,6 +19,56 @@ import XCTest
 
 final class CharacterExtensionsTests: XCTestCase {
 
+    func testIsEmoji() {
+        XCTAssert(Character("😂").isEmoji)
+        XCTAssertFalse(Character("j").isEmoji)
+    }
+
+    func testIsNumber() {
+        XCTAssert(Character("1").isNumber)
+        XCTAssertFalse(Character("s").isNumber)
+    }
+
+    func testIsLetter() {
+        XCTAssertTrue(Character("a").isLetter)
+        XCTAssertTrue(Character("B").isLetter)
+        XCTAssertFalse(Character("3").isLetter)
+        XCTAssertFalse(Character("-").isLetter)
+    }
+
+    func testIsLowercased() {
+        XCTAssert(Character("s").isLowercased)
+        XCTAssertFalse(Character("S").isLowercased)
+    }
+
+    func testIsUpercased() {
+        XCTAssert(Character("S").isUppercased)
+        XCTAssertFalse(Character("s").isUppercased)
+    }
+
+    func testIsWhiteSpace() {
+        XCTAssertTrue(Character(" ").isWhiteSpace)
+        XCTAssertFalse(Character("-").isWhiteSpace)
+    }
+
+    func testInt() {
+        XCTAssertNotNil(Character("1").int)
+        XCTAssertEqual(Character("1").int, 1)
+        XCTAssertNil(Character("s").int)
+    }
+
+    func testString() {
+        XCTAssertEqual(Character("s").string, String("s"))
+    }
+
+    func testUppercased() {
+        XCTAssertEqual(Character("s").uppercased, Character("S"))
+    }
+
+    func testLowercased() {
+        XCTAssertEqual(Character("S").lowercased, Character("s"))
+    }
+
 	func testOperators() {
 		let sLetter = Character("s")
 		XCTAssertEqual(sLetter * 5, "sssss")
@@ -29,56 +79,6 @@ final class CharacterExtensionsTests: XCTestCase {
 
 		XCTAssertEqual(sLetter * -5, "")
 		XCTAssertEqual(-5 * sLetter, "")
-	}
-
-	func testIsEmoji() {
-		XCTAssert(Character("😂").isEmoji)
-		XCTAssertFalse(Character("j").isEmoji)
-	}
-
-	func testIsNumber() {
-		XCTAssert(Character("1").isNumber)
-		XCTAssertFalse(Character("s").isNumber)
-	}
-
-	func testIsLetter() {
-		XCTAssertTrue(Character("a").isLetter)
-		XCTAssertTrue(Character("B").isLetter)
-		XCTAssertFalse(Character("3").isLetter)
-		XCTAssertFalse(Character("-").isLetter)
-	}
-
-	func testIsWhiteSpace() {
-		XCTAssertTrue(Character(" ").isWhiteSpace)
-		XCTAssertFalse(Character("-").isWhiteSpace)
-	}
-
-	func testInt() {
-		XCTAssertNotNil(Character("1").int)
-		XCTAssertEqual(Character("1").int, 1)
-		XCTAssertNil(Character("s").int)
-	}
-
-	func testString() {
-		XCTAssertEqual(Character("s").string, String("s"))
-	}
-
-	func testIsUpercased() {
-		XCTAssert(Character("S").isUppercased)
-		XCTAssertFalse(Character("s").isUppercased)
-	}
-
-	func testIsLowercased() {
-		XCTAssert(Character("s").isLowercased)
-		XCTAssertFalse(Character("S").isLowercased)
-	}
-
-	func testUppercased() {
-		XCTAssertEqual(Character("s").uppercased, Character("S"))
-	}
-
-	func testLowercased() {
-		XCTAssertEqual(Character("S").lowercased, Character("s"))
 	}
 
 }
